+++ ThemeDecorator/ThemeDecorator.js
@@ -3,6 +3,7 @@
  *
  * @module agate/ThemeDecorator
  * @exports ThemeDecorator
+ * @exports ThemeContext
  */
 
 import {setRootId} from '@enact/core/dispatcher';
@@ -22,6 +23,39 @@ import Skinnable from '../Skinnable';
 
 import screenTypes from './screenTypes.json';
 import css from './ThemeDecorator.module.less';
+
+const ThemeContext = React.createContext(null);
+
+const defaultColors = {
+	carbon: {
+		accent: '#8fd43a',
+		highlight: '#6abe0b'
+	},
+	cobalt: {
+		accent: '#8c81ff',
+		highlight: '#ffffff'
+	},
+	copper: {
+		accent: '#a47d66',
+		highlight: '#ffffff'
+	},
+	electro: {
+		accent: '#0359f0',
+		highlight: '#ff8100'
+	},
+	gallium: {
+		accent: '#8b7efe',
+		highlight: '#e16253'
+	},
+	silicon: {
+		accent: '#f1304f',
+		highlight: '#9e00d8'
+	},
+	titanium: {
+		accent: '#a6a6a6',
+		highlight: '#2a48ca'
+	}
+};
 
 /**
  * Default config for {@link agate/ThemeDecorator.ThemeDecorator}.
@@ -229,6 +263,7 @@ const CustomizableSkinStyle = kind({
 const ThemeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 	// TODO: Document props passable to hoc ()
 	const {customSkin, float, i18n, noAutoFocus, overlay, ri, skin, spotlight, disableFullscreen, rootId} = config;
+	const defaultSkin = 'gallium';
 
 	const bgClassName = classnames(
 		'enact-fit',
@@ -242,7 +277,7 @@ const ThemeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 	if (ri) App = ResolutionDecorator(ri, App);
 	if (i18n) App = I18nDecorator({sync: true}, App);
 	if (spotlight) App = SpotlightRootDecorator({noAutoFocus}, App);
-	if (skin) App = Skinnable({defaultSkin: 'gallium'}, App);
+	if (skin) App = Skinnable({defaultSkin}, App);
 
 	// add webOS-specific key maps
 	addAll({
@@ -257,7 +292,7 @@ const ThemeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 	const Decorator = class extends React.Component {
 		static displayName = 'ThemeDecorator';
 
-		static propTypes = {
+		static propTypes = /** @lends agate/ThemeDecorator.ThemeDecorator.prototype */ {
 			/**
 			 * A custom accent color, as a hex string.
 			 *
@@ -273,11 +308,20 @@ const ThemeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			 * @type {String}
 			 * @public
 			 */
-			highlight: PropTypes.string
+			highlight: PropTypes.string,
+
+			/**
+			 * The name of the skin a component should use to render itself.
+			 *
+			 * @type {String}
+			 * @public
+			 */
+			skin: PropTypes.string
 		};
 
 		render () {
-			const {accent, className, highlight, ...rest} = this.props;
+			const currentSkin = this.props.skin || defaultSkin;
+			const {accent = defaultColors[currentSkin].accent, className, highlight = defaultColors[currentSkin].highlight, ...rest} = this.props;
 			const customizableSkinClassName = 'agate-customized-skin';
 
 			const allClassNames = classnames(
@@ -292,10 +336,10 @@ const ThemeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			);
 
 			return (
-				<React.Fragment>
+				<ThemeContext.Provider value={{accent, highlight}}>
 					{customSkin ? <CustomizableSkinStyle className={customizableSkinClassName} accent={accent} highlight={highlight} /> : null}
-					<App {...rest} className={allClassNames} />
-				</React.Fragment>
+					<App {...rest} accent={accent} highlight={highlight} className={allClassNames} />
+				</ThemeContext.Provider>
 			);
 		}
 	};
@@ -305,5 +349,6 @@ const ThemeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 
 export default ThemeDecorator;
 export {
-	ThemeDecorator
+	ThemeDecorator,
+	ThemeContext
 };