+++ WindDirectionControl/WindDirectionControl.js
@@ -11,6 +11,7 @@
  */
 
 import kind from '@enact/core/kind';
+import {extractAriaProps} from '@enact/core/util';
 import Changeable from '@enact/ui/Changeable';
 import PropTypes from 'prop-types';
 import compose from 'ramda/src/compose';
@@ -39,6 +40,14 @@ const WindDirectionControlBase = kind({
 	name: 'WindDirectionControlBase',
 
 	propTypes: /** @lends agate/WindDirectionControl.WindDirectionControlBase.prototype */ {
+		/**
+		 * Whether or not the component is in a disabled state.
+		 *
+		 * @type {Boolean}
+		 * @public
+		 */
+		disabled: PropTypes.bool,
+
 		/**
 		 * Called when value is changed.
 		 *
@@ -70,10 +79,14 @@ const WindDirectionControlBase = kind({
 		airDirectionIcon: ({value}) => (value.toLowerCase())
 	},
 
-	render: ({airDirectionIcon, onChange, value, ...rest}) => {
+	render: ({airDirectionIcon, disabled, onChange, value, ...rest}) => {
+		const ariaProps = extractAriaProps(rest);
+
 		return (
 			<div {...rest}>
 				<ArcPicker
+					{...ariaProps}
+					disabled={disabled}
 					endAngle={210}
 					onChange={onChange}
 					slotCenter={